Accept interfaces changes only in discovery/error
By our design we accept changes to node interfaces only if node in
DISCOVERY and ERROR states. There are few reasons why we're doing so,
including one that IP addresses are assigned to bridges, not to NICs
and we ignore it in order to be able to redeploy nodes.

However, there's one case when node status is *artificially* changing
to ERROR state, implicitly allowing to accept interfaces changes. The
case is:

* Admin (PXE) interface is down on a node.
* nailgun-agent sends wrong node's IP address that will be used to
  check whether it belongs to Admin (PXE) network or not.
* If it doesn't belong, change node status to ERROR.
* Since the status is ERROR, changes to interfaces are allowed.
* Move admin network to another interface, which is wrong since
  physical connection wasn't changed.

The commit introduces additional condition when the check for belonging
to Admin (PXE) network is allowed.

Change-Id: I17e87e27d846921d6f0da535b9446e716449db95
Closes-Bug: #1532823
(cherry picked from commit 5925c52524cf4b1a540c021d96ef3e521bc71228)