Add processing of user defined VIPs by VIP assigning procedure
This patch implements processing of user supplied VIPs (already assigned).

In case VIP defined by user has non-empty namespace and check for
belonging to the network has not passed - error will be raised as
Nailgun cannot process user defined data in this situation. If user
defined VIP has empty namespace attribute it should be preserved as is
without check for belonging to the network.

In other cases (assuming the ip is within network's ranges)
VIP also should be preserved.

More details can be found in the related blueprint.

Change-Id: I3051c0d6c31e6cf58fb3c6893b2c768292de4787
Blueprint: allow-any-vip