Fix random failure on syncdb()
Nailgun tests fail time to time trying to execute syncdb(). This happens
because our tests keep dirty state after their execution, and some fake
releases may have 'roles_metadata = null' which brokes assumption of
our 9.0 migration.

It's better to mark this (and other) column as not-nullable in database,
however a lot of tests violate that constraint and require fixing.

So the best option is to fix migration in order to unblock other patches
on CI.

Change-Id: I072ebf2564d8d1d9312d2bd41a383a18587636a4
Closes-Bug: #1546566