Support feature advanced-config in task based deployment
The Advanced Config  allows to re-execute some tasks
in case of openstack configs was modified.
The flag 'refresh_on' of task uses to filter tasks
those should be reexecuted on modifying config.
Also added support for flag '*' that means reexecute
task in case if any of configs was modified.

Implements: blueprint enable-task-based-deployment
Change-Id: I0b3822c1c35d3dad6e7c3df4bb4a200bd63268ed