Fix Node.reset_to_discover method
Setting Node's attribute `online` manually is a bad idea. It's the
asssassind area of responsibility to set this attribute. If you do it
manually you must be sure that node really offline. If you don't check
it's may be the cause of some strange bugs such as bug/1440081. This
patch changes Node.reset_to_discover method so that it leave online
attribute the same and let assasind do the work.

Change-Id: I25b2e6ae5694652cd3606cc99fa5cd8542f73005
Closes-Bug: #1440081