Send cluster id for pre provision task obviously
Orchestrator implements a method called `calculate_cluster_id', which
extracts cluster id from the log file name that is supplied in the `cmd'
field of the parameters hash from pre_provision property of provisioning
info. We should send cluster id in a separate field (`cluster_id')
instead of inferring it from the string representation of the command.

Change-Id: Ia33810525707fa63985be019096dc9e78c81c812
Partial-Bug: #1449512