Robust generating of routes
'netgroups[ngname]' is replaced with 'netgroups.get(ngname, {})' as
in other places in serializer.
Problem was the 'br-prv' endpoint that didn't have IP in case of VLAN
segmentation. The difference between 'netgroup_mapping' and 'netgroups'
exists only in case of using network template. Otherwise, absence of IP
for 'br-prv' did not lead to this issue.

Also, there is a small refactoring of serialization tests was made here.

Conflicts:
	nailgun/nailgun/test/integration/test_orchestrator_serializer_90.py

Change-Id: I67616409017631023e707e533cb8e2faab943b34
Closes-Bug: #1556917