Do not include Admin network twice
Admin network was included into networks list twice in case of using
multiple node groups. Because of that the allocation algorithm had
an assumption that Admin network was not allocated to NIC and allocated
it twice, first time - to the right NIC, second - to the first NIC always.
The issue arrived when node was in non-default node group and was booted
not from the first NIC.
It is fixed now.

Conflicts:
	nailgun/nailgun/network/manager.py

Closes-Bug: #1557580
Change-Id: I64a9ecdad1e68e4f0819940f0266699c394e9caf