Added policy to format text w/o exception for method traverse
The method traverse can be used for formatting task parameters,
but it raises exception in case if format fails. The most shell tasks
contains symbols '{', '}' and python string format tries to format such
string. This change allows to specify the safe_formatter and
leave argument as is if format fails.

This patch also improves error handling for traverse, the default
text formatter added information about text that cannot be formatted.

Change-Id: I2d398d7a27966d842812d20ad1f5e7c5ecbff676
Related-bug: #1614401