Allow to use master role in conjunction with other
In some cases it is usefull to specify master role in conjuction with
other roles. for example if task should be execute on all nodes
including master node, it is usefull to write ['master', '/.*/']
The '*' and '/.*/' doest not include master node, the master role
should be specified explicitly.
Also allows to use string in role field, if there is only one role.

Change-Id: Icfdacd7b59dca7f5cfa09c9ee0695a5e8393bfa8
Implements: blueprint computable-task-fields-yaql