Fix check for cross-referring self-dependencies
This fix fixes an issue when the code of cross-reference
checking was too strict and checked not combination of
"task_name", "node_id",
but
only "node_id" instead.
This leads to the situation when cross-depends role: .*
does not include the task on the same node even
when task names of the required and of the requiring
tasks are different.

E.g. such dependency:

with nodes [1,2]
task: a
cross-depends:
   name: b

on node_id 1

should have requires: (b,1), (b,2)

but had requires: (b,2)

due to this bug

Change-Id: Ic8b6f86533670fe477cc68db6a0d932ee8501cc3
Closes-bug: #1559968