Do not print unhandled exception if in production
This change removes the printing of unhandled exceptions when not in
development mode. We should never return the unhandled exception
information when not in development mode.

Logging nailgun exceptions
Now there is a special "processor" devoted to this

Change-Id: I820f5682d0fac292722a9f205965516771333a4d
Closes-Bug: #1585160
(cherry picked from commit c61fbc616fc6763c358ad70dac37341f65ddcb19)