Remove legacy graph checker
Granular deployment has a checker for graph cycles which gets triggered
even for task-based deployment. It is really hard to refactor things
to make it run only for role-based deployment. As role-based deployment
is going to be deprecated and there is already another networkx/dag.py
checker in orchestrator_graph.py it seems ok to remove this checker from
all the places.

Change-Id: If081ce7a462bc5ced9459ba619d26aecf22f4916
Closes-bug: #1560555