Add exception handling when NIC with given name is missing in db
In case NIC names in 'nic_mapping' section of the network template do
not correspond to actually existing NICs of node, to which networks must
be assigned, the error will occur; as lookup in db by name is performed,
result of such lookup will be None, so any attempt to get attributes
from it raises AttributeError.

This change adds handling of the situation by adding of custom exception
and returning proper http response to user who tries to upload network
template.

Change-Id: Ia9d7264d3885805f91862e408f8059c4e5eed1d6
Partial-Bug: #1539115