Nailgun Fixes for smart configuration changes support
Simply unhardcode the handler for $.configuration namespace
to allow a user to modify anything he wants in this namespace.
The list of modifyable sections on '$.configuration' namespace of astute.yaml
should not be hardcoded - this makes no sense as support of these pieces
is implemented on fuel library and plugins side. Hardcoding the list
will make it impossible for a plugin developer to reuse configuration
changes feature to support configuration options for plugin.

DocImpact

Partial-bug: #1580522

Change-Id: I4eb18461c0a4f337b649abfcd099259965290ff5