Update interface_properties only when required
Root cause of the issue is a race condition that takes place
when processing of two API requests intersects.

These requests are:
update node info from nailgun-agent and adding the same node
into an environment.

interface_properties.dpdk section is updated on both requests.
So, it was possible that request from nailgun-agent that comes
during the processing of adding node into environment overwrites
dpdk.available property with wrong value (value that was there
before adding the node into environment). It is because the changes
made during addition of node were not committed before start of
processing the update request from nailgun-agent.

Now, update request from nailgun-agent will not overwrite this value
in such situation because no DB update will be performed when
no changes are detected.

Change-Id: Ica23cb2cfd3439d5bb4d51afa64095cf8256f6b6
Closes-Bug: #1564102