Cannot execute only selected tasks
It should be possible to select some tasks and execute
only the selected tasks instead of all the tasks.
But the current behavior does not allow to execute only the
selected tasks. This is due to the fact that the variable
conataining the selected tasks is overridden to contain
all the tasks before the execution. That variable should be
overridden only if the initial value was None.
This commit will fix the issue.

Change-Id: If3538e0fbd7ad4c77f9170f3c2f45bc0f987f765
Closes-Bug: #1570818