Add bootable disk node attribute
Adding this attridute to node allows user to choose bootable disk
from Fuel UI or CLi.

This change is very important for deployments with multipath. Mainly
it is because, order of disks in UI can change easly in this case,
so we need introduce to user a possibility to choose bootable disk
explicitly.

Change-Id: I22ffe9104d2ec5a6598d496691fffa0087111070
Partial-Bug: #1567450