Remove 'nodes' from deployment 9.0 serializer output
Now serialized deployment info for environments >= 9.0 does not contain
'nodes' field; it is removed from the output of the serializer.

Unit and integration tests are updated. In order to support (and get to
state where they more or less reflect actual production
environment) execution of tests utilizing the serializer, separate
fixture deployment tasks has been created. The only difference with fixture for
other release versions is that 'upload_nodes_info' task is skipped. Since the
fixture is uploaded on release creation; method that
supplies test environment with proper fixture in accordance to release
version being used is added. However this approach introduces a drawback
such that now separate fixtures must be introduced for each new release.
This issue is closely connected with the bug [1] and is know issue with
Nailgun testing infrastructure.

[1]: https://bugs.launchpad.net/fuel/+bug/1517874

Change-Id: Ibdb2cf85732c0293314298e6a342b071ece18cc3
Closes-Bug: #1531128