Moving snapshots to /var/log/dump
A partition used for /var is not big enough and, since snapshot contains
mostly logs anyway, it makes sense to move it to /var/log
Also do not inject master node host creds - it doesn't make sense to
login to the localhost via ssh and make a copy of all data instead of
just symlinking it.
Also fixes incorrect excludes.

Change-Id: I2e1a60bd71955b2c607f46d2cde60daa49811be7
Partial-Bug: #1543491