Message in receiver requeued on deadlock
Seems, that we have regression in the Nailgun code that brings deadlocks.
When deadlock occurs on message processing in the receiverd we mark message
as acknowledged.
As partial fix we requeue message on deadlock in the receiverd.

Change-Id: I408fcadb9068154a4d355ee1ab206c42ecad7c64
Partial-Bug: #1578218
(cherry picked from commit e42b6318be1ee6d7275fa91152dee64b54b06546)