Networks configuration from UI is now available for stopped nodes
So it is possible to update network configuration for node in
'stopped' state using NodeCollectionHandler

Network config update via NodeAgentHandler is still restricted but
it is still possible to update metadata from this handler.

'is agent' request flag is now added automatically if not persist
when update is launched via NodeAgentHandler.

Change-Id: I5f41c74553b4227cd968900fb121d26ff2029cda
Partial-Bug: #1581527