Allow YAQL expressions in task's dependencies
Since Fuel 9.0 we have a task-based deployment turned on by default.
That means our tasks are executing simultaneously on several nodes, and
old static approach for specifying dependencies doesn't work anymore.

In order to avoid dependency loops and make possible to rely on
different tasks based on some condition, we've got to be able to
calculate `requires`, `required_for`, `cross-depends` and
`cross-depended-by` attributes on fly.

Fortunately, we already have traversal mechanism that evaluates YAQL
expressions almost everywhere. So all we need to do is to remove
validation limitation and allow to upload deployment tasks with YAQL
expressions in those fields.

Change-Id: I393dfb06f9e25ecd3ca79fddc84104a8cf993094
Partial-Bug: #1541309