Keep only actually required attributes in cluster context
We do not have clear boundary between node and cluster attributes,
and we cannot use whole master node state as cluster state because
it contains the node specific attributes too. The temporary
solution is add only list of nodes and product version to cluster
attributes, because only this attributes are needed as old state for
nodes, that do not have previous state.

Change-Id: I12ac64048c47a1afea03dc07b4c8ba3ba2c3ce2e
Closes-Bug: 1582247