Fuel controls S3 API/Keystone integration of Ceph RadosGW.
This change is a part of the broader effort on letting Fuel to control
the S3 API/Keystone integration of Ceph RadosGW. The overall idea has
been described in the specs/10.0/s3-keystone-integration.rst document
that is a part of the fuel-specs repo. It is expected that the UI change
will be complemented by a corresponding modification of fuel-library.

Change-Id: I6badbd8bfb73b4c50112b016dc53ee3017de40ec
Related-Bug: #1446704
Related-Bug: #1540426