Expose configuration to skip gateway ping check
We currently allow for someone to bypass the public gateway ping check
on the controllers for the public_vip monitor via hiera. But this is
hidden and not well documented. This change aims to expose this to the
deployer in a more obvious way.

Change-Id: Ic244e907fb9db2c020cf590e96626b015cbe0773
Closes-Bug: #1524640
(cherry picked from commit 36515bbbc13a582c92abd46d67843c0939d8a738)