Clear primary roles during upgrade of nodes
This patch clears primary_roles of a node during its reassignment as a
part of an upgrade. Without this fix it is possible to have several
primary roles in a seed cluster during the upgrade. This race condition
leads to some failed deployments during the upgrade. As a result primary
roles have to be cleared during the reassignment of nodes. That allows
to assign primary roles on appropriate nodes rely on the internal logic.

Related-bug: 1599837
Change-Id: Iae5f3090cfbfd1e6033b94720694a94477562328
(cherry picked from commit aca06b5296b3c5dc7f3bac3b1d61e3d3ec76c686)