Clear primary roles during upgrade of nodes
This patch clears primary_roles of a node during its reassignment as a
part of an upgrade. Without this fix during upgrade of a cluster it was
possible to have several primary roles in a seed cluster. This race
condition leads to some failed deployment during the upgrade. As a
result primary roles have to be cleared during a reassignment of nodes
that allows to assign primary roles on appropriate nodes rely the
internal logic.

Related-bug: 1599837
Change-Id: Iae5f3090cfbfd1e6033b94720694a94477562328