Resolve tasks by node labels
Tasks resolution is now done based on node labels instead
of roles. For backwards compatability all node roles are
added as labels during serialization. Serialization will
continue to work as before if no labels are specified.

Tasks can specify a 'labels' attribute which will be used
to match them to nodes. If 'labels' does not exist it will
fall back to using 'roles'.

Change-Id: I390580146048b6e00ec5c42d0adf995a4cff9167