Add ability to tag nodes
Tags are associated with each node and are freeform. They are
managed by the API endpoint at /tags.

A 'tags' attribute has been added to each role in 'roles_metadata'.
Initially all non-controller roles will only have a tag of their own
role name. This will allow existing tasks which do not have tags
associated with them to work correctly. In the abscence of tags a
task's roles will be used to determine which nodes it will run on.

Change-Id: I390580146048b6e00ec5c42d0adf995a4cff9167
Implements: blueprint role-decomposition