remove response "validation"
It seems somebody wanted to implement response
validation, but now it's completely broken: it's done in
checked_data method, which takes data either from web input
or from data keyword, and second, there is only one
validate_response method which it empty.

So let's remove it now and if somebody will have spare
time in the future, he could implement it from scratch.

Change-Id: Ic60236fce11998d76354d3d2707d0260351ca063