Fixed updating node status in transaction resp
The transient statuses are not persited in DB because the status of node
should represent the current state of node, for example: node is
provisioned or node is deployed.  The synthetic states: like error,
deploying, etc. are used only for providing additional for user about
status of node at this moment.  This statuses can be calculated on
demand.  Also node.error_type has type 'String' instead of 'Enum' Also
added handling status  'deleted' in response, which means that node
should be deleted from cluster.

Change-Id: If5e79b9274f34e01d2b795491c23361c9050669d
Blueprint: graph-concept-extension