Track timestamps for start and end of transactions
In order to enable building a history of deployment transactions
for a custer it is necessary to keep track of start and end time
for them. This patch adds timestamp_start and timestamp_end attributes
to transactions, makes changes to the API serializers and introduces
appropriate database fields to store that information

I've added times only for 'deployment' task, which is the only
one, needed by new UI feature. At this point there is no single
point, where task switches to error or ready state and I'd have
to set in in many places.

Whem everything-is-a-graph approach will be fully implemented
it would be very easy. So let's not add unnecessary code now.

Backport of Ib206d75a8d2215f8fd6b3f89e209c6e86ed20d0f

DocImpact
Closes-bug: #1593753

Change-Id: Ib206d75a8d2215f8fd6b3f89e209c6e86ed20d0f