Add ability to get common or node part of context in lcm
The main purpose of this commit is to have an ability
to split configuration file astute.yaml into common
and node parts. Common part is huge and we will
dump it once and also there will be only one instance
of this data in RAM which saves a lot of memory when
you run deploy on many nodes (>100).

This patch adds two new variables to context, which can be
used in yaql_exp: $node and $common for node and common
parts of the context. Functions changed, changed_all,
changed_any, added, deleted don't work for these variables.

DocImpact
Change-Id: I56bf982652a5dc27882e4a401ca9ec124899fed7
Partial-Bug: #1596987