Remove un-needed filters in nodegroup tests
The default admin-pxe network has a (node)group of None as it is handled
specially, when added to a cluster. This prevents any other nodegroup from
having an pxe network subsection on the same network range, this behavior is
un-expected as it functions correctly when the non-default nodegroup is one
of the members in the test.

Change-Id: Iaa3037561d31c2d13892e1d796afc0865eec8a5d
Closes-bug: #1623098