Fix role validation in release PUT handler
Plugins' roles can be assigned to some nodes, so we can't
assume that release roles is be a superset of node roles
everywhere.

So check that there are no nodes in all clusters, which
have roles assigned, that we are trying top delete.

Forward-port of Ia10f7b0662866da0872863bba7f97c8ad361fcf4

Change-Id: I28c23bd239ca150832323c8994788cb3aeb40eef
Closes-Bug: #1625592