Fix role validation in release PUT handler
Plugins' roles can be assigned to some nodes, so we can't
assume that release roles is be a superset of node roles
everywhere.

So check that there are no nodes in all clusters, which
have roles assigned, that we are trying top delete.

Forward-port of 78fbccc9ac1fad1f05bfc9e68677d4ef24b1a0aa

Change-Id: I28c23bd239ca150832323c8994788cb3aeb40eef
Closes-Bug: #1625592