Avoid creating tasks with empty nodes list
For a granular deploy when a node is deleted and deployment is
triggered the list of nodes for deploy is empty and all remaining nodes
are in the list of affected nodes. This causes nonreexecutable tasks to
appear in the list of tasks with empty uids array.
This patch fixes the problem.

Change-Id: I7377c50df87d1f2448ac2bc775840738d535d7fd
Closes-Bug: #1628500
(cherry picked from commit 70ebe005390bfa475801dcba8b3dc1374996a994)