Fix referring to wrong cobbler profile in ubuntu release data
Apparently, release metadata contained referring to wrong cobbler
profile. That cobbler profile was used a while ago for native
installer purposes and became invalid since only IBP method of
provisioning remained.

The problem is that every node will be added to cobbler with
wrong profile and enabled netbooting. If provisioning failed for
some reasons, then the node won't boot because of stale broken links
to initrd and kernel.

`generator_arg` should reffer to valid name of cobbler profile for
 bootstrap image.

For 8.0 and newer, it should be 'ubuntu_bootstrap'
For 7.0 it should be 'bootstrap' (which stands for centos bootstrap).

Change-Id: I86c1cc2501281a85878ac584096a7d12dc4576da
Closes-Bug: #1626962