Reexecute tasks priorities calculation fixed
If we have the same task in deployement and in reexecution tasks
and have other tasks which requires the first one, then after
merging deployment and reexection tasks to the single list
priorities can be broken. For instance child task from the
reexection tasks can have priority higher than parent task in the
deployment tasks list. As result child task from the reexecutable
tasks will be merged into deplyment tasks list and tasks
priorites be broken.
As fix we should check priorites of tasks on merge of reexecutable
and deployment tasks list and use the highest priority in the
reslut tasks list.

Also fixed mistypes in comments and dockstrings.

Change-Id: Ic029973f673658c2ff00a5eb4561bdfcfa960cd6
Closes-Bug: #1644871