Do not lock nodes records for a long time
Before cluster serialization primary tags on nodes are updated. This
results in updated nodes record to be locked until the transaction where
serialization happens ends. When there is a lot of nodes this
transaction can be very long and, as a result, a lot of node updates
can be blocked until it ends. At some points this results in all workers
being busy blocking any REST call.

Change-Id: Ie5341290097bddce299bf5726e2607d150c23768
Closes-Bug: #1643024