Change node search and validation methods
- Remove searches by any MAC
- Add search by system uuid
- Allow duplicate MACs (except "mac" property)

The previous implementation used a fallback matching rule which
would match a node if any of its MACs are found in the database.
This is not reliable because MACs can sometimes be identical due to
hardware specifics or other reasons.

This change removes matching by any MAC with matching by system uuid,
which is reliably unique for both hardware and virtual systems,
and also constant for a given system unless it's motherboard is
replaced.

This change allows nodes to have identical MACs on iterfaces.
However, "mac" property still has to be unique.

User is responsible for ensuring that interfaces with identical MACs
(if any) are not assigned to the same networks, which will obviously
lead to deployment & connectivity issues.

This change adds a randomly generated system uuid to the default_metadata
in tests.

Closes-Bug: 1650646
Change-Id: I49855e2b5ff30559cb2385a03abb8cfdd6a5f1f2