Change node search fallback method to system uuid
The previous implementation used a fallback matching rule where
it would match a node if any of its MACs are identical to any
of the MACs already in DB. This is not reliable because MACs can
sometimes be identical due to hardware specifics or other reasons.

This change removes matching by any MAC with matching by system uuid,
which is reliably unique for both hardware and virtual systems,
and also constant for a given system unless it's motherboard is
replaced.

Change-Id: I49855e2b5ff30559cb2385a03abb8cfdd6a5f1f2
Closes-Bug: 1650646