Logs symlinks creation fixed
We had a race on the old symlink recreation. After old symlink
removing rsyslog creates a dir with old name and after that
symlink creation is failed due to 'File exists' error. As fix
we are stopping the rsyslog for removing old link as well as
it was done for removing old dir.

Change-Id: I7fa969c3495f7e15a3b9f44ee9eaccca4106d5c1
Closes-Bug: #1652771