Get rid of excessive flush
This flush is not needed anymore since there is dedicated flush just
before update_interfaces.
Excessive flush leads to two udpates instead of one and creates a
situation when deadlock with concurrent node updates is possible.
The deadlock, actualy, happens on scale. The concurrent transaction is
node progress update by the receiver.

Change-Id: Ib30193a5e299cdd2144d9e12b0f06c99e2914960
Closes-Bug: #1624230