Update node data only for specified statuses
When node is provisioning, deploying or removing we shouldn't update
node data. Repeatable updates causes deadlock when ShareLock is not
granted in the other transaction due to locks of the DB keys. So in
case of locking task, cluster, node in receiver and repeatable updates
of node in the API handler we had a deadlock.

Change-Id: I7addd5a9e256403df91e818d0efcefc347989789
Partial-Bug: #1624230