Populate context with transaction meta information
In order to be able to have tasks parameters serialized
based on the context of the particular transaction,
we need to have this info available within yaql interpreter.

E.g. for some of the provisioning tasks we need to have the
exact list of nodes that we have sent for the graph execution.

This commit adds a new YAQL context variable called 'meta' and
populates it with a list of nodes that are processed by the particular
subtransaction.

Change-Id: I2ac3a2bee8ffedb835bf64383495424acf63e4f9
Partial-bug: #1667006