Do not reset node to discovery state
We use info about nodes for user message and system
reset. In original code Astute should not send
any info about nodes if we run stop for task deployment
which used in deploy stage, but such behavior
create a problem which was fixed in #1590458.
It works because Astute use another action for provision
stop, but last changes also introduced provision as graph which
used task deployment to run provision.

Now Astute run stop deploy actions for provision as for deploy and
info about nodes start to used as reference to action:
reset node status to discovery.

This fix remove code which reset nodes to discovery stage in case
of provisioning. It can be removed safely because it created
in times when Cobbler was used as main provision tools (before
image provision).

Change-Id: I32da2ccce11b22378f58759703fc4a56e31fd993
Closes-Bug: #1672964