Wait till reset task will return ready status
For some reason we wait for running or pending task when we test
environment's resets. And then for some reason asserting that the
task is ready. As a result, we get sporadic py27 failures if
computation resources are limites.

Proposed patch ensures that we wait for correct status of the
task to be executed.

Change-Id: I0fe41f41ca1776cccb033b93795eb8cd016ad3dc
Closes-bug: #1677570