Validate that plugins exist before sync procedure
There are two problems in Nailgun if we try to sync all plugins
that actually not even exist (neither in db nor on filesystem).
1. By default 'plugins' directory in '/var/www/nailgun/' does not
exist that leads to an error if we try to sync plugins before
install them.
2. If 'plugins' directory exists, but doesn't contain any plugins
and there are no any installed plugins in db, Nailgun reports
confusing message "Plugins were successfully synchronized."
So, this patch adds extra validation procedure that allows sync
plugins only if they exist in Nailgun (either db or filesystem)

Change-Id: Id06d2da1174bbed512c26cff22f51de91100c66d
Closes-Bug: 1677627